Streamlined pybind11 examples using dpctl4pybind11.hpp
